+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 2.1.2 (????-??-??)
 * Refresh button in admin toolbar for admin as well as front-end
 * Fix for IE9 sources hack in `pilau_responsive_image()`
+* Hack to handle symlinked CMB2 in local dev
 
 ## 2.1.1 (2015-06-21)
 * `pilau_string_of_one_liners_to_array()`